fix(Refactor): Changes class/file names + created appopriate directories
