+++ Bot/Cogs/Commands/QuoteCommands.py
