+++ Bot/Cogs/Managers/QuoteManager.py
